refactor/Changed maker equipment props
